% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{GetRedcapToken}
\alias{GetRedcapToken}
\title{Supply a user REDCap token. End users should edit this file.}
\usage{
GetRedcapToken()
}
\description{
Supply a user REDCap token. End users should edit this file.
}
