% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary-file-tools.R
\name{HardCopyRightThere}
\alias{HardCopyRightThere}
\title{Write a csv file somewhere within or at the end of a pipe.}
\usage{
HardCopyRightThere(d, f = "hardcopy.csv", addtime = T)
}
\arguments{
\item{f}{a string e.g. "file.csv"}

\item{addtime}{a logical indicating whether to add a timestamp}
}
\description{
Write a csv file somewhere within or at the end of a pipe.
}
