% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{name}
\alias{name}
\title{A string representing the final path component, excluding the drive and root, if any}
\usage{
name(f = "hardcopy.csv")
}
\description{
A string representing the final path component, excluding the drive and root, if any
}
