% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary-file-tools.R
\name{ReadHardCopy}
\alias{ReadHardCopy}
\title{Read a file from HardCopyRightThere}
\usage{
ReadHardCopy(f = "hardcopy.csv")
}
\description{
The user supplies a candidate filename as a string, the function sorts all files in the directory matching that filename and loads the first in sorted order.
Note that if HardCopyRightThere has supplied a timestamped file, this will read the first match without a timestamp, or, barring that, the most recent timestamp.
}
