% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadNrmAverages}
\alias{ReadNrmAverages}
\title{Read the average data from the nrm file. note that in some cases a large number of empty columns are generated; only extract the first three.}
\usage{
ReadNrmAverages(nrm_file)
}
\description{
Read the average data from the nrm file. note that in some cases a large number of empty columns are generated; only extract the first three.
}
