% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadNrmTimeseries}
\alias{ReadNrmTimeseries}
\title{Read the timeseries data from the nrm file. this produces raw data of 100x46 doubles.
to scale, you should add to database. to get done, you should do in memory.
keeping the kinematic marker data will definitely need a database.}
\usage{
ReadNrmTimeseries(nrm_file)
}
\description{
Read the timeseries data from the nrm file. this produces raw data of 100x46 doubles.
to scale, you should add to database. to get done, you should do in memory.
keeping the kinematic marker data will definitely need a database.
}
