% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadPersonalData}
\alias{ReadPersonalData}
\title{Read subject data (height, weight, MRN) from personal.dat files generated by Orthotrack}
\usage{
ReadPersonalData(file.name)

ReadPersonalData(file.name)
}
\description{
Read subject data (height, weight, MRN) from personal.dat files generated by Orthotrack

Read anthropemetric data from personal.dat.
}
