% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadRedcapReport}
\alias{ReadRedcapReport}
\title{Read data from a REDCap report. Default arguments depend on specific project.}
\usage{
ReadRedcapReport(
  token = GetRedcapToken(),
  url = "https://redcap.emory.edu/api/",
  report_id = "29228"
)
}
\description{
Read data from a REDCap report. Default arguments depend on specific project.
}
