% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadTrbCoord}
\alias{ReadTrbCoord}
\title{Read the estimated joint centers from the .trbcoord files.}
\usage{
ReadTrbCoord(file.name)
}
\description{
Read the estimated joint centers from the .trbcoord files.
}
