% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadTrcData}
\alias{ReadTrcData}
\title{Read in the first 60 (or n.markers) columns of data from a .trc file.}
\usage{
ReadTrcData(file.name, n.markers = 60)
}
\description{
Read in the first 60 (or n.markers) columns of data from a .trc file.
}
