% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadTrcHeader}
\alias{ReadTrcHeader}
\title{Read trc file header.
Note that on read-in, backslashes in filenames are rendered as "\\" this is the backslash character in r, no need to attempt to remove.}
\usage{
ReadTrcHeader(file.name)
}
\description{
Read trc file header.
Note that on read-in, backslashes in filenames are rendered as "\\" this is the backslash character in r, no need to attempt to remove.
}
