% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadTrcVariableNames}
\alias{ReadTrcVariableNames}
\title{Read in the first 60 (or n.markers) variable names.}
\usage{
ReadTrcVariableNames(file.name, n.markers = 60)
}
\description{
Read in the first 60 (or n.markers) variable names.
}
