% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep12-file-tools.R
\name{UnixizeMotionPath}
\alias{UnixizeMotionPath}
\title{Replace common directory locations like Z:\ with the MotionAnalysis root directory}
\usage{
UnixizeMotionPath(x)
}
\description{
Replace common directory locations like Z:\ with the MotionAnalysis root directory
}
