% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary-file-tools.R
\name{ReadHardCopy}
\alias{ReadHardCopy}
\title{Read the most recent csv with a name spec from HardCopyRightThere}
\usage{
ReadHardCopy(f = "hardcopy.csv")
}
\description{
Read the most recent csv with a name spec from HardCopyRightThere
}
