% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{ReadRedcapReport}
\alias{ReadRedcapReport}
\title{Read data from a REDCap report.
Arguments are as supplied in API playground.}
\usage{
ReadRedcapReport(
  token,
  url = "https://redcap.emory.edu/api/",
  report_id = "29228",
  format = "csv",
  csvDelimiter = "",
  rawOrLabel = "raw",
  rawOrLabelHeaders = "raw",
  exportCheckboxLabel = "false",
  returnFormat = "json"
)
}
\description{
The API token should be stored in the user's .Renviron file.
}
